requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiogram-media-group==0.5.1
aiohttp==3.8.4
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
config==0.5.1
environs==9.5.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
marshmallow==3.20.1
multidict==6.0.4
openai==0.27.4
packaging==23.2
psycopg2-binary
pydantic==1.10.7
pydantic_core==2.14.5
python-dotenv==1.0.0
regex==2023.3.23
requests==2.28.2
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions~=4.6.0
urllib3==1.26.15
yarl==1.8.2
YooMoney==0.1.0
